requirements.txt
scikit-learn
mne
pyriemann
matplotlib
seaborn
h5py
pandas
pyyaml
coloredlogs
sphinx-gallery
sphinx-bootstrap_theme
numpydoc
pillow
requests
m2r2
patool
pyunpack
wfdb
